+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - `--time=mtime` and `--time=modification`.
+- `--time=birth` and `--time=creation`, but now it falls back to `--time=mtime` or `--time=modification`.
 
 ## 0.6.1.0 -- 2023-05-26